atualizar start.sh

+++ backend/mock/start.sh
@@ -2,7 +2,11 @@
 
 current_path=$(pwd)
 
-npm install json-server@0.17.4 jsonwebtoken cors
+package_file="$current_path/package.json"
+
+if [[ ! -f "$package_file" ]]; then
+    npm install json-server@0.17.4 jsonwebtoken cors
+fi
 
 cp "$current_path/db.seed.json" "$current_path/db.json"
